RENAME: fake_http -> charlatan
it turns out there's a defunct pub.dev package called fake_http already
:(

so this is a proposed alternative name.

a charlatan is a faker or a fraud. it's one of those cutesy names and i
always have mixed feelings about them, but it definitely helps with
avoiding naming conflicts in a language without namespaces like dart.