feat(response-customization): add CharlatanHttpResponse
add the ability to customize the entire response within the response
definition's response builder callback.

this allows runtime configuration of the response statusCode, headers,
and body; instead of just the body.

it's a little cute, but we're returning Object? from the response
builder still since we don't have the ability to use type unions.

we could probably make it so that you always have to return a response
object, but it's so many more characters to type in the case where you
just want the response to be simple :shrug: