feat: `whenMatch` - for more advanced matching needs
sometimes, you want to define your own matching logic and now you can.
if you just want to match when a certain request parameter is present or
if the http method is GET and the first letter of the path is "z". who
knows what you'll need? the sky is now the limit.

breaking change: removed CharlatanHttpRequest#pathParameters. it is no
longer possible to support this attribute of the request type because
the request is provided in the matching phase when we don't know what
the path template is and whether its a match yet.