feat! add charlatanResponse helper, add type-safety
fixes #24

this changeset adds a new `charlatanResponse` helper function that
returns a `CharlatanResponseBuilder` with a default status code.

this new helper function replaces the previous approach of returning an
`Object?` from the `CharlatanResponseBuilder` type. This makes it safer
to work with in your codebase but is a breaking change.

To migrate to this version, replace any usages of directly returning
anything other than `CharlatanHttpResponse` in `when*` methods with a
a call to `charlatanResponse` or an explicit `CharlatanHttpResponse`
instance.